rename feature folder
