+++ src/features/auth/signUp/ui/SignUpForm.tsx
@@ -2,7 +2,7 @@ import githubIcon from '@/shared/assets/icons/github-icon.svg'
 import googleIcon from '@/shared/assets/icons/google-icon.svg'
 import { PRIVACY_POLICY, SIGN_IN, TERMS_OF_SERVICE } from '@/shared/config/router'
 import { Button } from '@/shared/ui/button/button'
-import Card from '@/shared/ui/card/Card'
+import { Card } from '@/shared/ui/card'
 import { Input } from '@/shared/ui/input/input'
 import Image from 'next/image'
 import Link from 'next/link'
